getting changes from real azure

+++ src/azure-cli-testsdk/azure/cli/testsdk/preparers.py
@@ -5,6 +5,7 @@
 
 import inspect
 import functools
+import os
 
 from .base import ScenarioTest, execute
 from .exceptions import CliTestError
@@ -21,17 +22,16 @@ class AbstractPreparer(object):
         self.resource_moniker = None
         self.resource_random_name = None
         self.test_class_instance = None
+        self.live_test = False
 
     def __call__(self, fn):
         def _preparer_wrapper(test_class_instance, **kwargs):
-            if not isinstance(test_class_instance, ScenarioTest):
-                raise CliTestError('The preparer decorator can be only used on the methods of '
-                                   'class derived from {}'.format(ScenarioTest.__name__))
+            self.live_test = not isinstance(test_class_instance, ScenarioTest)
             self.test_class_instance = test_class_instance
 
-            if test_class_instance.in_recording:
+            if self.live_test or test_class_instance.in_recording:
                 resource_name = self.random_name
-                if isinstance(self, RecordingProcessor):
+                if not self.live_test and isinstance(self, RecordingProcessor):
                     test_class_instance.recording_processors.append(self)
             else:
                 resource_name = self.moniker
@@ -82,7 +82,13 @@ class AbstractPreparer(object):
 class SingleValueReplacer(RecordingProcessor):
     # pylint: disable=no-member
     def process_request(self, request):
-        request.uri = request.uri.replace(self.random_name, self.moniker)
+        from six.moves.urllib_parse import quote_plus  # pylint: disable=import-error
+        if self.random_name in request.uri:
+            request.uri = request.uri.replace(self.random_name, self.moniker)
+        elif quote_plus(self.random_name) in request.uri:
+            request.uri = request.uri.replace(quote_plus(self.random_name),
+                                              quote_plus(self.moniker))
+
         if request.body:
             body = str(request.body)
             if self.random_name in body:
@@ -104,20 +110,32 @@ class SingleValueReplacer(RecordingProcessor):
 # Resource Group Preparer and its shorthand decorator
 
 class ResourceGroupPreparer(AbstractPreparer, SingleValueReplacer):
-    def __init__(self, name_prefix='clitest.rg', parameter_name='resource_group',
-                 parameter_name_for_location='resource_group_location', location='westus'):
-        super(ResourceGroupPreparer, self).__init__(name_prefix, 90)
+    def __init__(self, name_prefix='clitest.rg',  # pylint: disable=too-many-arguments
+                 parameter_name='resource_group',
+                 parameter_name_for_location='resource_group_location', location='westus',
+                 dev_setting_name='AZURE_CLI_TEST_DEV_RESOURCE_GROUP_NAME',
+                 dev_setting_location='AZURE_CLI_TEST_DEV_RESOURCE_GROUP_LOCATION',
+                 random_name_length=75):
+        super(ResourceGroupPreparer, self).__init__(name_prefix, random_name_length)
         self.location = location
         self.parameter_name = parameter_name
         self.parameter_name_for_location = parameter_name_for_location
 
+        self.dev_setting_name = os.environ.get(dev_setting_name, None)
+        self.dev_setting_location = os.environ.get(dev_setting_location, location)
+
     def create_resource(self, name, **kwargs):
-        template = 'az group create --location {} --name {} --tag use=az-test'
-        execute(template.format(self.location, name))
-        return {self.parameter_name: name, self.parameter_name_for_location: self.location}
+        if self.dev_setting_name:
+            return {self.parameter_name: self.dev_setting_name,
+                    self.parameter_name_for_location: self.dev_setting_location}
+        else:
+            template = 'az group create --location {} --name {} --tag use=az-test'
+            execute(template.format(self.location, name))
+            return {self.parameter_name: name, self.parameter_name_for_location: self.location}
 
     def remove_resource(self, name, **kwargs):
-        execute('az group delete --name {} --yes --no-wait'.format(name))
+        if not self.dev_setting_name:
+            execute('az group delete --name {} --yes --no-wait'.format(name))
 
 
 # Storage Account Preparer and its shorthand decorator
@@ -126,7 +144,7 @@ class StorageAccountPreparer(AbstractPreparer, SingleValueReplacer):
     def __init__(self,  # pylint: disable=too-many-arguments
                  name_prefix='clitest', sku='Standard_LRS', location='westus',
                  parameter_name='storage_account', resource_group_parameter_name='resource_group',
-                 skip_delete=True):
+                 skip_delete=True, dev_setting_name='AZURE_CLI_TEST_DEV_STORAGE_ACCOUNT_NAME'):
         super(StorageAccountPreparer, self).__init__(name_prefix, 24)
         self.location = location
         self.sku = sku
@@ -134,14 +152,19 @@ class StorageAccountPreparer(AbstractPreparer, SingleValueReplacer):
         self.skip_delete = skip_delete
         self.parameter_name = parameter_name
 
+        self.dev_setting_name = os.environ.get(dev_setting_name, None)
+
     def create_resource(self, name, **kwargs):
-        group = self._get_resource_group(**kwargs)
-        template = 'az storage account create -n {} -g {} -l {} --sku {}'
-        execute(template.format(name, group, self.location, self.sku))
-        return {self.parameter_name: name}
+        if not self.dev_setting_name:
+            group = self._get_resource_group(**kwargs)
+            template = 'az storage account create -n {} -g {} -l {} --sku {}'
+            execute(template.format(name, group, self.location, self.sku))
+            return {self.parameter_name: name}
+        else:
+            return {self.parameter_name: self.dev_setting_name}
 
     def remove_resource(self, name, **kwargs):
-        if not self.skip_delete:
+        if not self.skip_delete and not self.dev_setting_name:
             group = self._get_resource_group(**kwargs)
             execute('az storage account delete -n {} -g {} --yes'.format(name, group))
 
@@ -153,6 +176,36 @@ class StorageAccountPreparer(AbstractPreparer, SingleValueReplacer):
                        'decorator @{} in front of this storage account preparer.'
             raise CliTestError(template.format(ResourceGroupPreparer.__name__,
                                                self.resource_group_parameter_name))
+
+
+# Role based access control service principal preparer
+
+class RoleBasedServicePrincipalPreparer(AbstractPreparer, SingleValueReplacer):
+    def __init__(self, name_prefix='http://clitest',  # pylint: disable=too-many-arguments
+                 skip_assignment=True, parameter_name='sp_name', parameter_password='sp_password',
+                 dev_setting_sp_name='AZURE_CLI_TEST_DEV_SP_NAME',
+                 dev_setting_sp_password='AZURE_CLI_TEST_DEV_SP_PASSWORD'):
+        super(RoleBasedServicePrincipalPreparer, self).__init__(name_prefix, 24)
+        self.skip_assignment = skip_assignment
+        self.result = {}
+        self.parameter_name = parameter_name
+        self.parameter_password = parameter_password
+        self.dev_setting_sp_name = os.environ.get(dev_setting_sp_name, None)
+        self.dev_setting_sp_password = os.environ.get(dev_setting_sp_password, None)
+
+    def create_resource(self, name, **kwargs):
+        if not self.dev_setting_sp_name:
+            command = 'az ad sp create-for-rbac -n {}{}'\
+                      .format(name, ' --skip-assignment' if self.skip_assignment else '')
+            self.result = execute(command).get_output_in_json()
+            return {self.parameter_name: name, self.parameter_password: self.result['password']}
+        else:
+            return {self.parameter_name: self.dev_setting_sp_name,
+                    self.parameter_password: self.dev_setting_sp_password}
+
+    def remove_resource(self, name, **kwargs):
+        if not self.dev_setting_sp_name:
+            execute('az ad sp delete --id {}'.format(self.result['appId']))
 
 
 # Utility
